fix: arithmetic overflow
bench: 2172435

+++ engine/src/history_table.rs
@@ -27,9 +27,9 @@ impl HistoryTable {
         let clamped_bonus = bonus.clamp(-Score::MAX_HISTORY, Score::MAX_HISTORY);
         let current_value = self.table[side as usize][piece as usize][square as usize];
         // history gravity formula
-        let new_value = current_value.0 + clamped_bonus
-            - current_value.0 * clamped_bonus.abs() / Score::MAX_HISTORY;
-        self.table[side as usize][piece as usize][square as usize] = Score(new_value);
+        let new_value = current_value.0 as i32 + clamped_bonus as i32
+            - current_value.0 as i32 * clamped_bonus.abs() as i32 / Score::MAX_HISTORY as i32;
+        self.table[side as usize][piece as usize][square as usize] = Score(new_value as i16);
     }
 
     pub(crate) fn clear(&mut self) {
